backend/requirements.txt
Flask==2.0.1
pdfplumber==0.5.28
scikit-learn==0.24.2
numpy==1.22.4
pandas==1.3.5
nltk==3.6.7
joblib==1.1.0
requests==2.26.0
Werkzeug==2.0.2
PyMuPDF==1.19.6
spacy==3.2.0
gunicorn==20.1.0
